moved auth functions to Navigator.js, moved sensitive info to .env
